DXNES: use moving_threshold based on AR process
The moving threshold proposed in the original paper assumes that
when convergence stagnates, evolutionary path ~N(0,I), and
so E|path| ~ E(Chi(dims)). But in stagnation phase uZ ~ N(0,I) and so
evolutionary path: path[t+1] = a*path[k]+b*uZ is in fact autoregressive
process with white noise. In that case
E|path|^2=Var(|path|)=b^2*dims/(1-a^2).

The old threshold was too sensitive, the system was considered moving
when it was stagnating. It resulted in the uncontrolled growth of sigma.